+++ examples/basic_keystore.rs
@@ -3,13 +3,14 @@
 use db_keystore::{DbKeyStore, DbKeyStoreConfig};
 use keyring_core::{Result, api::CredentialStoreApi};
 use std::collections::HashMap;
+use zeroize::Zeroizing;
 
 const SERVICE: &str = "basic_example";
 
 fn main() -> Result<()> {
     // initialize keystore in default location
     let config = DbKeyStoreConfig::default();
-    let store = DbKeyStore::new(&config)?;
+    let store = DbKeyStore::new(config)?;
     println!("using store at {}", store.path());
 
     // set password for user alice
@@ -22,8 +23,8 @@ fn main() -> Result<()> {
 
     // retrieve a password
     let results = store.search(&HashMap::from([("service", SERVICE), ("user", "alice")]))?;
-    let password = results.get(0).unwrap().get_password()?;
-    assert_eq!(password, "dromomeryx");
+    let password = Zeroizing::new(results.first().unwrap().get_password()?);
+    assert_eq!(password.as_str(), "dromomeryx");
 
     // cleanup: search for all users using regex and delete
     for entry in store.search(&HashMap::from([